minor change - moved LoginFlag type to common package
